PIM-4252: move csv reader to connector component and depreciate the old one
