+++ spec/Pim/Component/Connector/Reader/File/CsvReaderSpec.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace spec\Pim\Component\Connector\Reader\File;
+
+use PhpSpec\ObjectBehavior;
+
+class CsvReaderSpec extends ObjectBehavior
+{
+    function it_is_configurable()
+    {
+        $this->setFilePath('/path/to/file/');
+        $this->setDelimiter(';');
+        $this->setEnclosure('-');
+        $this->setEscape('\\');
+        $this->setUploadAllowed(true);
+
+        $this->getFilePath()->shouldReturn('/path/to/file/');
+        $this->getDelimiter()->shouldReturn(';');
+        $this->getEnclosure()->shouldReturn('-');
+        $this->getEscape()->shouldReturn('\\');
+        $this->isUploadAllowed()->shouldReturn(true);
+    }
+
+    function it_gives_configuration_field()
+    {
+        $this->getConfigurationFields()->shouldReturn([
+            'filePath' => [
+                'options' => [
+                    'label' => 'pim_base_connector.import.filePath.label',
+                    'help'  => 'pim_base_connector.import.filePath.help'
+                ]
+            ],
+            'uploadAllowed' => [
+                'type'    => 'switch',
+                'options' => [
+                    'label' => 'pim_base_connector.import.uploadAllowed.label',
+                    'help'  => 'pim_base_connector.import.uploadAllowed.help'
+                ]
+            ],
+            'delimiter' => [
+                'options' => [
+                    'label' => 'pim_base_connector.import.delimiter.label',
+                    'help'  => 'pim_base_connector.import.delimiter.help'
+                ]
+            ],
+            'enclosure' => [
+                'options' => [
+                    'label' => 'pim_base_connector.import.enclosure.label',
+                    'help'  => 'pim_base_connector.import.enclosure.help'
+                ]
+            ],
+            'escape' => [
+                'options' => [
+                    'label' => 'pim_base_connector.import.escape.label',
+                    'help'  => 'pim_base_connector.import.escape.help'
+                ]
+            ],
+        ]);
+    }
+}
